Plugin - Product Search - Refactored some package naming, added integration tests
