+++ software/plugins/external-item-search/src/integrationTest/java/tech/ebp/oqm/plugin/extItemSearch/BootTestIT.java
@@ -1,0 +1,10 @@
+package tech.ebp.oqm.plugin.extItemSearch;
+
+import io.quarkus.test.junit.QuarkusTest;
+import org.junit.jupiter.api.Test;
+
+import static io.restassured.RestAssured.given;
+
+@QuarkusTest
+class BootTestIT extends BootTest {
+}
