--- software/plugins/external-item-search/src/main/java/tech/ebp/openQuarterMaster/plugin/productSearch/service/searchServices/ItemSearchService.java
@@ -1,11 +1,0 @@
-package tech.ebp.openQuarterMaster.plugin.productSearch.service.searchServices;
-
-
-import tech.ebp.openQuarterMaster.plugin.productSearch.model.ExtItemLookupProviderInfo;
-
-public abstract class ItemSearchService {
-	
-	public abstract ExtItemLookupProviderInfo getProviderInfo();
-	
-	public abstract boolean isEnabled();
-}
