+++ software/plugins/external-item-search/src/main/java/tech/ebp/oqm/plugin/extItemSearch/service/searchServices/ItemSearchService.java
@@ -1,0 +1,11 @@
+package tech.ebp.oqm.plugin.extItemSearch.service.searchServices;
+
+
+import tech.ebp.oqm.plugin.extItemSearch.model.ExtItemLookupProviderInfo;
+
+public abstract class ItemSearchService {
+	
+	public abstract ExtItemLookupProviderInfo getProviderInfo();
+	
+	public abstract boolean isEnabled();
+}
