+++ software/plugins/external-item-search/src/main/java/tech/ebp/oqm/plugin/extItemSearch/service/searchServices/webPage/GenericWebProductScrapeService.java
@@ -1,0 +1,58 @@
+package tech.ebp.oqm.plugin.extItemSearch.service.searchServices.webPage;
+
+import io.opentelemetry.instrumentation.annotations.WithSpan;
+import jakarta.enterprise.context.ApplicationScoped;
+import lombok.NoArgsConstructor;
+import lombok.SneakyThrows;
+import lombok.extern.slf4j.Slf4j;
+import org.jsoup.nodes.Document;
+import tech.ebp.oqm.plugin.extItemSearch.model.ExtItemLookupProviderInfo;
+import tech.ebp.oqm.plugin.extItemSearch.model.ExtItemLookupResult;
+
+import java.net.URL;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+
+@ApplicationScoped
+@Slf4j
+@NoArgsConstructor
+public class GenericWebProductScrapeService extends WebPageProductScrapeService {
+	
+	@SneakyThrows
+	@Override
+	public ExtItemLookupProviderInfo getProviderInfo() {
+		//TODO:: add to application.yaml?
+		return ExtItemLookupProviderInfo.builder()
+				   .displayName("Generic")
+				   .description("Generic page scraper; will attempt to scrape any page given with best-effort.")
+				   .acceptsContributions(false)
+				   .cost("Free")
+				   .enabled(this.isEnabled())
+										.build();
+	}
+	
+	@WithSpan
+	@Override
+	protected ExtItemLookupResult scrapePageContent(Document webPageContent) {
+		ExtItemLookupResult.Builder<?, ?> resultBuilder = this.getInitialBuilder(webPageContent);
+		
+		Map<String, String> atts = new HashMap<>();
+		
+		//TODO:: this
+		
+		
+		return resultBuilder.build();
+	}
+	
+	
+	@Override
+	public List<String> supportedHosts() {
+		return null;
+	}
+	
+	@Override
+	public boolean canParsePage(URL url) {
+		return true;
+	}
+}
