+++ software/plugins/external-item-search/src/native-test/java/tech/ebp/openQuarterMaster/BootIT.java
@@ -1,6 +1,7 @@
 package tech.ebp.openQuarterMaster;
 
 import io.quarkus.test.junit.QuarkusIntegrationTest;
+import tech.ebp.oqm.plugin.extItemSearch.BootTest;
 
 @QuarkusIntegrationTest
 class BootIT extends BootTest {
