--- software/plugins/external-item-search/src/test/java/tech/ebp/openQuarterMaster/BootTest.java
@@ -1,19 +1,0 @@
-package tech.ebp.openQuarterMaster;
-
-import io.quarkus.test.junit.QuarkusTest;
-import org.junit.jupiter.api.Test;
-
-import static io.restassured.RestAssured.given;
-import static org.hamcrest.CoreMatchers.is;
-
-@QuarkusTest
-class BootTest {
-    @Test
-    void testHelloEndpoint() {
-        given()
-          .when().get("/api/v1/providers")
-          .then()
-             .statusCode(200);
-    }
-
-}
