+++ software/plugins/external-item-search/src/test/java/tech/ebp/oqm/plugin/extItemSearch/BootTest.java
@@ -1,0 +1,19 @@
+package tech.ebp.oqm.plugin.extItemSearch;
+
+import io.quarkus.test.junit.QuarkusTest;
+import org.junit.jupiter.api.Test;
+
+import static io.restassured.RestAssured.given;
+import static org.hamcrest.CoreMatchers.is;
+
+@QuarkusTest
+class BootTest {
+    @Test
+    void testHelloEndpoint() {
+        given()
+          .when().get("/api/v1/providers")
+          .then()
+             .statusCode(200);
+    }
+
+}
